refactor: move Query responses to a more general package
