+++ services/identity-service/src/main/java/io/ciphernance/identity/application/query/account/byownerid/GetAccountByOwnerIdQuery.java
@@ -1,7 +1,7 @@
 package io.ciphernance.identity.application.query.account.byownerid;
 
 import io.ciphernance.identity.application.mediator.Query;
-import io.ciphernance.identity.application.query.account.status.GetAccountResponse;
+import io.ciphernance.identity.application.query.account.GetAccountResponse;
 
 import java.util.UUID;
