+++ services/identity-service/src/main/java/io/ciphernance/identity/application/query/account/status/GetAccountStatusQuery.java
@@ -1,6 +1,7 @@
 package io.ciphernance.identity.application.query.account.status;
 
 import io.ciphernance.identity.application.mediator.Query;
+import io.ciphernance.identity.application.query.account.GetAccountResponse;
 
 import java.util.UUID;
