+++ services/identity-service/src/main/java/io/ciphernance/identity/application/query/user/profile/GetUserProfileHandler.java
@@ -2,6 +2,7 @@ package io.ciphernance.identity.application.query.user.profile;
 
 import io.ciphernance.identity.application.exception.user.UserNotFoundException;
 import io.ciphernance.identity.application.mediator.QueryHandler;
+import io.ciphernance.identity.application.query.user.GetUserProfileResponse;
 import io.ciphernance.identity.domain.model.User;
 import io.ciphernance.identity.domain.port.UserRepositoryPort;
 import org.springframework.stereotype.Component;
