+++ services/identity-service/src/main/java/io/ciphernance/identity/application/query/user/profile/GetUserProfileQuery.java
@@ -1,6 +1,7 @@
 package io.ciphernance.identity.application.query.user.profile;
 
 import io.ciphernance.identity.application.mediator.Query;
+import io.ciphernance.identity.application.query.user.GetUserProfileResponse;
 
 import java.util.UUID;
